requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.13.1
